edit stories
