+++ stories/components/QSelect/Default.js
@@ -1,0 +1,91 @@
+const options = [
+  {
+    value: { id: 'value1' },
+    label: 'Option 1'
+  },
+  {
+    value: { id: 'value2' },
+    label: 'Option 2'
+  },
+  {
+    value: { id: 'value3' },
+    label:
+      'Lorem Ipsum is simply dummy text of the printing and typesetting industry.'
+  },
+  {
+    value: { id: 'value4' },
+    label: 'Option 4',
+    disabled: true
+  },
+  {
+    value: { id: 'value5' },
+    label: 'Sed sit amet nibh consequat, pellentesque arcu ut, congue lorem.',
+    disabled: true
+  },
+  {
+    value: { id: 'value6' },
+    label: 'Option 6'
+  }
+];
+
+export default (_, { argTypes }) => ({
+  props: Object.keys(argTypes).filter(arg => arg !== 'value'),
+
+  data() {
+    return {
+      remoteLoading: false,
+      value: null,
+      options
+    };
+  },
+
+  watch: {
+    multiple() {
+      this.value = null;
+    },
+
+    remote(value) {
+      if (!value) this.options = options;
+    }
+  },
+
+  methods: {
+    handleSearch(query) {
+      if (!this.remote) return;
+
+      this.remoteLoading = true;
+
+      setTimeout(() => {
+        this.remoteLoading = false;
+
+        if (query !== '') {
+          this.options = options.filter(item => {
+            return item.label.toLowerCase().includes(query.toLowerCase());
+          });
+        } else {
+          this.options = options;
+        }
+      }, 2000);
+    }
+  },
+
+  template: `
+      <div style="width: 304px;">
+        <q-select
+          v-model="value"
+          v-bind="$props"
+          :loading="loading || remoteLoading"
+          @search="handleSearch"
+          placeholder="Pick an option"
+        >
+          <q-option
+            v-for="item in options"
+            :key="item.value.id"
+            :label="item.label"
+            :value="item.value"
+            :disabled="item.disabled"
+          />
+        </q-select>
+      </div>
+    `
+});
