+++ stories/components/QSelect/Multiple.js
@@ -1,0 +1,15 @@
+import Template from './Default';
+
+const Multiple = Template.bind({});
+
+Multiple.args = {
+  appendToBody: false,
+  multiple: true,
+  filterable: true,
+  valueKey: 'id',
+  allowCreate: false,
+  collapseTags: false,
+  clearable: true
+};
+
+export default Multiple;
